Rename file service to fileList service
Fixes #16
